Explicit upfront OS/platform checking
The install scripts perform some upfront checks before attempting
to download binaries for Health Center core. This will alert a
user if a binary is not expected to exist for a given OS/arch/version.
This is better than failing later trying to download the library,
because it would be the same error as you see if the download failed
for some other reason.

Previously we listed a few operating systems that we did not support.
In this commit we change to an explicit list of platforms we expect
to be available for the current Health Center core agent version.

Now that we no longer download the appmetrics.node binaries when
we install from github (it is compiled from source), I have removed
the unused code related to that.

NOTE: We publish a slightly different NPM to npmjs.org which downloads
      all binaries and compiles nothing from source (at present),
      in case the end user does not have a compiler installed.