Make use of nan require resilient to failure
gyp will run the require line even if a different
value of nandir is passed on the commandline.

In the case where there is no nan on the node module
path, but a valid nandir has been passed in, this means
the error thrown by require('nan') will prevent
the build from working.

This change catches the error and instead populates the
nandir with the error message. If a nandir is passed
in then the error message will be replaced with the value
passed.

If no nandir is passed then the build failure message should
include the nan require failure message.