Fix npm test failure on dev versions
The tests assumed the build identifier would look like one from a
full release and did not recognise a build identifier from a dev
release.

For example, this build id for release version 1.0.7 was recognised:
    1.0.7.201602110606

...but this build id for development version 1.0.8-dev.0 was not:
    1.0.8-dev.0.201604071706

This commit changes the test to recognise the build identifer from
a dev release.

Related: #186