Fix recursive require of appmetrics
A package such as appmetrics cannot require itself by name
(`require('appmetrics')`), its unresolvable. If and only if appmetrics happens
to be installed as a dependency of a package will that work.

An example of when it doesn't is its own unit tests. If you clone appmetrics
and run its unit tests, they fail:

	ibm/appmetrics (adapter % u=) % npm test

	> appmetrics@1.0.9-dev.0 test /home/sam/w/ibm/appmetrics
	> node tests/test_runner.js

	module.js:339
	    throw err;
	    ^

	Error: Cannot find module 'appmetrics'